feat(license): refactor editor flow
