+++ src/components/Editor/SettingsDialog/SetCover/Selector.tsx
