+++ src/components/Editor/SettingsDialog/SetCover/styles.css
