--- src/components/Editor/Sidebar/Box/styles.css
@@ -1,12 +1,0 @@
-.box {
-  @mixin border-grey;
-
-  padding: var(--spacing-base) 0;
-  border-radius: var(--spacing-x-tight);
-
-  & header {
-    @mixin flex-center-space-between;
-
-    padding: 0 var(--spacing-base);
-  }
-}
