--- src/components/Editor/Sidebar/Collapsable/styles.css
@@ -1,11 +1,0 @@
-section {
-  @mixin border-bottom-grey;
-
-  padding-bottom: var(--spacing-loose);
-  margin-bottom: var(--spacing-loose);
-  color: var(--color-grey);
-
-  &.expanded {
-    color: var(--color-black);
-  }
-}
