--- src/components/Editor/Sidebar/Collection/styles.css
@@ -1,11 +1,0 @@
-ul {
-  padding-top: var(--spacing-base);
-
-  & li {
-    @mixin border-top-grey;
-
-    &:last-child {
-      margin-bottom: calc(var(--spacing-base) * -1);
-    }
-  }
-}
