--- src/components/Editor/Sidebar/Cover/styles.css
@@ -1,19 +1,0 @@
-.container {
-  padding: var(--spacing-base) var(--spacing-base) 0;
-}
-
-.cover {
-  position: relative;
-  width: 100%;
-  height: 0;
-  padding-top: 56.25%;
-  overflow: hidden;
-  background: var(--color-grey-lighter);
-  border-radius: var(--spacing-xxx-tight);
-
-  & img {
-    @mixin object-fit-cover;
-
-    background-color: var(--color-grey-lighter);
-  }
-}
