--- src/components/Editor/Sidebar/Management/styles.css
@@ -1,6 +1,0 @@
-.container {
-  @mixin border-top-grey;
-
-  padding: var(--spacing-base) var(--spacing-base) 0;
-  margin-top: var(--spacing-base);
-}
