--- src/components/Editor/Sidebar/index.tsx
@@ -1,13 +1,0 @@
-import Collection from './Collection'
-import Cover from './Cover'
-import Management from './Management'
-import Tags from './Tags'
-
-const Sidebar = {
-  Cover,
-  Tags,
-  Collection,
-  Management,
-}
-
-export default Sidebar
