+++ src/views/Circle/Profile/AddCircleArticle/Dialog/index.tsx
@@ -9,10 +9,8 @@ import {
   useMutation,
   ViewerContext,
 } from '~/components'
-import {
-  SearchSelectDialog,
-  SearchSelectNode,
-} from '~/components/Dialogs/SearchSelectDialog'
+import { SearchSelectDialog } from '~/components/Dialogs/SearchSelectDialog'
+import { SearchSelectNode } from '~/components/Forms/SearchSelectForm'
 import PUT_CIRCLE_ARTICLES from '~/components/GQL/mutations/putCircleArticles'
 
 import { ADD_TOAST, REFETCH_CIRCLE_DETAIL_ARTICLES } from '~/common/enums'
