+++ src/views/Me/DraftDetail/index.tsx
@@ -8,7 +8,6 @@ import {
   Layout,
   Spinner,
   Throw404,
-  useResponsive,
   useRoute,
 } from '~/components'
 import { QueryError, useMutation } from '~/components/GQL'
@@ -17,12 +16,10 @@ import UPLOAD_FILE from '~/components/GQL/mutations/uploadFile'
 import { ASSET_TYPE, ENTITY_TYPE } from '~/common/enums'
 import { stripHtml } from '~/common/utils'
 
-import BottomBar from './BottomBar'
 import { DRAFT_DETAIL, SET_CONTENT } from './gql'
-import PublishButton from './PublishButton'
 import PublishState from './PublishState'
 import SaveStatus from './SaveStatus'
-import Sidebar from './Sidebar'
+import SettingsButton from './SettingsButton'
 
 import { SingleFileUpload } from '~/components/GQL/mutations/__generated__/SingleFileUpload'
 import { DraftDetailQuery } from './__generated__/DraftDetailQuery'
@@ -34,7 +31,6 @@ const Editor = dynamic(() => import('~/components/Editor/Article'), {
 })
 
 const DraftDetail = () => {
-  const isLargeUp = useResponsive('lg-up')
   const { getQuery } = useRoute()
   const id = getQuery('draftId')
 
@@ -144,16 +140,19 @@ const DraftDetail = () => {
   }
 
   return (
-    <Layout.Main
-      aside={<Sidebar draft={draft} ownCircles={ownCircles} />}
-      inEditor
-    >
+    <Layout.Main>
       <Layout.Header
         left={<Layout.Header.BackButton />}
         right={
           <>
             <SaveStatus status={saveStatus} />
-            {draft && <PublishButton disabled={!publishable} />}
+            {draft && (
+              <SettingsButton
+                draft={draft}
+                ownCircles={ownCircles}
+                disabled={!publishable}
+              />
+            )}
           </>
         }
       />
@@ -171,8 +170,6 @@ const DraftDetail = () => {
       <Layout.Spacing>
         <Editor draft={draft} update={update} upload={upload} />
       </Layout.Spacing>
-
-      {!isLargeUp && <BottomBar draft={draft} ownCircles={ownCircles} />}
     </Layout.Main>
   )
 }
